Core: allow clients to connect without a client certificate or key, if the server allows it.  To enable, add the following to the profile:
    setenv CLIENT_CERT 0

This is necessary to resolve an ambiguity when the profile
contains no client certificate or key, because otherwise
the client app can't know whether an external certificate/key
pair should be obtained from the Keychain, or whether the
server actually doesn't require a client certificate/key.
If the above directive is set to 1 or absent, the app will
assume that an external certificate/key pair should be obtained
from the Keychain

The option is given as a "setenv" to avoid breaking other
OpenVPN clients that might not recognize it.

----

Also, made subtle change to autologin determination, so that
community external PKI profiles will work properly with
autologin or userlogin based on the presence or absence of
auth-user-pass.