OpenVPN 1.0.4 build 113 (iOS)
Core: Added forceAesCbcCiphersuites config flag.  When enabled,
the TLS implementation will not set a minimum TLS version for
peer negotiation (even if tls-version-min is specified), but
will instead force one of these two ciphersuites:

1. TLS_DHE_RSA_WITH_AES_256_CBC_SHA, or
2. TLS_DHE_RSA_WITH_AES_128_CBC_SHA

When forceAesCbcCiphersuites is disabled, no explicit set of
ciphersuites will be forced, but the TLS minimum version will
be set to 1.0 or (or higher if tls-version-min is specified).

Also added iOS preference and cli.cpp option to control the
forceAesCbcCiphersuites flag.