Added "Exclude Routes Emulation" feature for both IPv4 and IPv6.
This feature is needed by Android because it lacks a native
VPN API method for excluding routes.

If redirect-gateway is enabled and exclude routes are present,
such as:

  route 54.215.128.71 255.255.255.255 net_gateway

the client will emulate the excluded route(s) by adding routes
that encompass the entire IPv4/v6 address space EXCEPT for the
excluded route.  These routes will be used for redirect-gateway
instead of the standard 0.0.0.0/0 and ::0/0.