Still having problem with SSL_pending() on SSLv23 raising an error instead of doing the sane thing and just returning 0.
To fix this, we actually override some of the OpenSSL SSLv23 methods here,
in particular the ssl_pending method.  We want ssl_pending
to return 0 until the SSL negotiation establishes the
actual method.  The default OpenSSL SSLv23 ssl_pending method
(ssl_undefined_const_function) triggers an OpenSSL error condition
which is not what we want.