Implemented tcp-queue-limit parameter for OpenVPN 3 client, with a default value of 64 (this is the same parameter name and default from OpenVPN 2 server).
tcp-queue-limit sets a maximum size limit on the outgoing
TCP packet queue and will drop packets and increment the
TCP_OVERFLOW counter if the limit would be exceeded.

Previously, the limit was hardcoded at 1024 packets and
would result in a disconnect/reconnect if exceeded rather
than merely dropping the packet.