Remotelist: pass meaningful port value to resolve::async_resolve()
OVPN3-13

Passing a static empty string as port to async_resolve() was
systematically triggering an exception when std::stoi() was
trying to parse it. On top of that, the port_ member of the
Preresolve object would not be instantiated with the right port value.

Fix it by passing the parsed port value coming from the config as
argument.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>