tls-crypt-v2: implement client key parser and renderer
The client key used for tls-crypt-v2 is stored in PEM format
and it is made up by the actual client key (Kc) and the
wrapped client key (WKc). The latter is an envelope containing
Kc (and some optional metadata) encrypted by the server with
its own key. It is sent upon connection to allow the server to
extract the actual Kc to be used for the tls-crypt session.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>