tls-crypt-v2: introduce CONTROL_HARD_RESET_V3 packet type
In order to allow a server to distinguish between an old session
and a new tls-crypt-v2 session reset, introduce a new packet type
to be used when initializing the latter.

The difference with V2 is that the "wrapped client key" (WKc) is
appended to this packet by the client. WKc contains the client
key and it can be unwrapped only by the server.

Failure to unwrap or authenticate this packet will result in
no reply from the server.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>