tls-crypt: make HMAC API more generic
In order to make the HMAC API in the TLCrypt class
re-usable by tls-crypt-v2, avoid using the hard-coded
header size of the standard tls-crypt packet.

Instead, let the caller specify the header size via
argument.

Note that the header size is also expected to be the offset
where it is possible to find the Authentication tag to be
used during the packet authentication.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>