[OC-42] Android: specify API level on command line
since the introduction of the unified headers, the API level
has to be specified on the command line, otherwise a development
API level (1000) will be used and will mess up older Android
releases.

On top of that, circumvent a bug in pthread.h by defining __LP32__.

This problem was causing crashes on Android 5 as it wasn't able to
provide libc functions expected by the precompiled ovpn3-core.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>