[OVPN3-105] ParseClientConfig: avoid crash when not all key material is provided
Some profiles are expected to not contain each of the key/ca/cert
material and sometimes an encrypted key might be provided.

In these cases the ParseClientConfig object would trigger an exception
as the underlying SSLConfig object was not able to be fully
instantiated.

However this is expected with some profiles, therefore catch such
exceptions and continue the execution normally.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>