[OVPN3-129] build-system: let scripts download dependencies
With this change a dep build script will now download
the related tarball automatically if not already present.

This way, we ensure that the core is built with the dep
package version specified in lib-versions.

After finding or downloading a tarball, its checksum is
computed and compared with the one in lib-versions to
ensure that the file is the expected one.

This logic has been applied to asio, mbedtls and lz4.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>