Estblishing a stable branch
This branch is targeted for consumers of the library, where we will do
release handling.  This branch should contain reasonably well tested and
stable code only.  Features under development and testing should only be
worked on outside of this branch until it is ready and can be merged
into stable.

Since it has been quite some time since the version.hpp file was
updated, this change will now set version to 3.2 - to start this new
versioning regime.

The version numbering and release process is further described in
VersionNumbering.rst

Signed-off-by: David Sommerseth <davids@openvpn.net>