proto.hpp/tls-crypt: fix access to ACK IDs during packet validation
Due to a typ0 in the validate_tls_crypt() function,
ack_skip() is invoked with the not-yet decrypted packet
as argument instead of the decrypted one.

This leads to buffer exceptions, becuse ack_skip() will
read a bogus ACK array length instead of the proper value.

This bug may lead to renegotiations issues on clients
due to a forced rejection of soft-reset packets in
proto.hpp:control_net_recv().

Fix the issue by passing the proper packet buffer to
ack_skip().

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>