httpserv::client: add and invoke tcp_intercept() before consuming buffer
Child classes may want to operate on a received buffer, before it is
passed down the stack. This can be useful when the Proxy Protocol
parser wants to parse and wipe its header.

Make it a virtual method so that it can be overridden by child classes
where higher level logic are implemented.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>