[UCONNECT-868] When no network is present pause instead of stopping
The previous version of the client used to error out. This is
undesirable behaviour as the UI should normally only need to worry about
high level behaviour (connection desired vs not desired) and should not
be required to watch for the next network status change event to start
the client on that event.

This commit changes the behaviour of the client to enter a pause state
and waiting for the network to become available if it is started without
network.