[UCONNECT-1027] implement ResolveThread and ensure it is properly detachable
The new thread we create to perform the async DNS resolution must be
fully detachable. This is a strong requirement, because its parent (the
AsyncResolvable class) and the core itself may disappear by the time the
DNS resolution thread is ready to post the callback.

This situation can easily happen when the DNS resolution is hanging on a
non-working network, while the user has already terminated the core by
explicitly clicking 'disconnect' on the the UI.

Fix this issue by creating a ResolveThread class which can receive
a 'detach' signal from the parent when the latter is about to disappear.

The ResolveThread will then be able to understand that it was left alone
and will not post any callback.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>