[UCONNECT-1027] remotelist: create standalone object for resolve thread
The core might be asked to solve several hostnames and for this reason
it is racy to use one AsioWork object to coordinate them all.
The reason why it is racy is that a thread may likely be unsetting the
AsioWork object while we still have another ongoing resolution. This
will lead to the original problem of "the i/o queue is empty, core can
exit now".

To fix this, create a standalone ResolveThread class that is
instantiated once per resolve operation and that contains its private
AsioWork object.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>