[UCONNECT-1027] use one AsioWork object for the whole pre-resolve opertation
Using AsioWork directly inside a ResolveThread resulted in the
impossibility of the core to kill it, thus re-creating the original
problem of requiring the core thread to wait for the resolve to be
finished.

The original "i/o queue can be empty problem" appears only in case we
are performing a pre-resolve operation, because this is when there is
nothing else scheduled for the i/o reactor. Resolution operation within
tcp/udp/http clients do not really have this problem because they happen
in an already active context.

For this reason AsioWork can be moved out of the resolve thread and can
be used only during pre-resolve.
At the same time we want to retain control of the AsioWork object
because we HAVE TO kill it when the core is trying to stop. This
operation will ensure that the i/o reactor can be immediately
terminated.

For this reason, we move the AsioWork object into in the AsyncResolvable
class that is always reachable and allows the core to kill the Asiowork
object when needed (this was impossible when AsioWork was part of
ResolveThread).

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>