Minor change to Error::INACTIVE_TIMEOUT handler
Previously, the Error::INACTIVE_TIMEOUT handler in cliconnect.hpp
called graceful_stop() to ensure that the Explicit Exit Notify
message is sent to the server prior to disconnect.

But actually, the Explicit Exit Notify message is sent earlier
by the inactive_callback() method in cliproto.hpp, and the
graceful_stop() call in the Error::INACTIVE_TIMEOUT handler
is redundant and unnecessary.

So this patch changes the graceful_stop() call to stop().

Signed-off-by: James Yonan <james@openvpn.net>