openssl: Improve OpenSSLContext fencing against multiple declarations
OpenSSLContext::SSL::ssl23_method_client_ and
OpenSSLContext::SSL::ssl23_method_server_ are only defined if the
OPENSSL_VERSION_NUMBER is below 0x10100000L.  But this will still make
multiple compilation units fail, due to how the various #include lines
are ordered and used.  This can eventually lead to multiple declarations
of ssl23_method_client_ and ssl23_method_server_ - which happens to
coreUnitTests on RHEL-7.

We add an additional defined(OPENVPN_NO_EXTERN) test to be sure this is
really not declared unless externs are expected to be declared.

Signed-off-by: David Sommerseth <davids@openvpn.net>