openssl: Add --verify-x509-name support
This implements the --verify-x509-name support for builds with OpenSSL.
This change requires the x509_get_subject() to be extended to provide
the subject field in a newer format, which requires using a different
OpenSSL API.  Since we have code which might require the old format as
well, x509_get_subject() will default to use the old format.

The --verify-x509-name supports matching both against the full subject
line and the X.509 Certificate Common Name, which means we need to check
which mode is configured and extract the proper value before calling the
VerifyX509Name::verify() method.

Signed-off-by: David Sommerseth <davids@openvpn.net>