openssl: Move generic X.509 cert extraction code to a separate file
This moves generic OpenSSL related functions used to extract X.509
certificate information to an independent file which can be used more
freely.  This code does not contain anything specific to the
OpenSSLContext class group and works nicely as a separate unit.

This is needed to be able to add unit test code for the x509_get_*()
functions.

Signed-off-by: David Sommerseth <davids@openvpn.net>