cliopt.hpp: disable remote list bypass for agent-enabled build
When tun_persist is set, we do pre-resolve and install bypass routes
for all remotes to be able to reconnect in force-tunneling case and tunnel is broken.
This is sub-optimal, since DNS may change during VPN session and there is no need
to have bypass route for remote A when we are connected to remote B.

Following commits will add bypass route before opening tunnel.

Signed-off-by: Lev Stipakov <lev@openvpn.net>