Linux/TunMethods: Avoid adding IPv6 routes without an IPv6 config
In some situations, the local6 variable is nullptr but a default IPv6
route has been configured.  This causes a segfault later in the call
chain when add_del_route() is being called.

We already have avoid a similar situation with IPv4, so implement the
same kind of safe guard for IPv6:  If no local IPv6 address has been
configured, don't attempt to add IPv6 routes.

Signed-off-by: David Sommerseth <davids@openvpn.net>