Fix static allocation in memq to be class static instead
The bio_memq_internal::memq_method_type looks extremely similar to the
usage of the OpenSSLContext::SSL::ssl_data_index.

Both are declared static. But the difference is that bio_memq_internal
is a namespace and SSL is a class.

The namespace around static memq_method_type just decorates the names
and the variable is a static variable, which is very problematic in our
header only approach as it ends up being defined in every compilation
unit. Normally static variables are just inside a single c or cpp file
and not in a header.

Since all our header methods are declared inline these are weak global
symbols, so only one copy of methods survives and that one will only
access its "own" static variable. The compiler on the other hand might
have inlined other methods accessing the other copy of the static
variable.

The static variable inside a class on the  other hand creates non local
variable with global linkage so only copy of the variable ends up in the
finished program.

This commit changes the bio_memq_internal memspace into a class. This
class has only static methods and members and has a small overhead of
unused constructor/destructor but it makes the usage here identical to
the other usages of the pattern in OpenVPN3