Avoid maybe unintialized warning in IP::Addr
Since the memory layout of the IPAddr class has the field ver behind
the union of u.v4 and u.v6, the whole u is always guaranteed to be
valid and can be copied. This avoid the compiler warning that
u.v6[1] might be undefined.

Also initialise the union u by default.

Signed-off-by: Arne Schwabe <arne@openvpn.net>