Simplify HMAC initialised logic to check just ctx != nullptr
Before the OpenSSL 1.1 conversion, we used HMAC_CTX as a field and the
variable initalised to signal if it is initialised. Since it
got converted to a pointer with OpenSSL 1.1 we can remove the
initialised variable just check if ctx != nullptr if it is initialised.

HMAC_CTX_free is (like free()) also allowed on a nullptr.

This also fixes a ctx might not be initialised warning on Fedora 31

Signed-off-by: Arne Schwabe <arne@openvpn.net>