Add warnings for 64bit block cipher and for SHA1 signed certificates
Example with ovpncli:

EVENT: WARN Proto: Using a 64-bit block cipher that is vulnerable to the SWEET32 attack. Please inform your admin to upgrade to a stronger algorithm. Support for 64-bit block cipher will be dropped in the future.

VENT: WARN TLS: received certificate signed with SHA1. Please inform your admin to upgrade to a stronger algorithm. Support for SHA1 signatures will be dropped in the future

Signed-off-by: Arne Schwabe <arne@openvpn.net>