win/client/tunsetup.hpp: fix redirect-gw case for local remote
The case when server is in local network and pushes
redirect-gw wasn't properly handled - gw.defined() is false,
but gw.localroute() is true. We threw an exception
because we weren't able to get a gateway for bypass route.

Since in "server in local network" case bypass route
is not needed, fix it by skipping code branch which
checks for the gateway and installs bypass route.

Signed-off-by: Lev Stipakov <lev@openvpn.net>