Add PluggableTransports protocols
Add a Factory, Transport, and Connection interface to support Pluggable Transports.

Factory: Generates a Transport object. OpenVPNClient has a virtual method that should be over-ridden by the implementer.
Transport: Is able to "dial" an endpoint and can generate a Connection.
Connection: Can be used to read and write data to the transport. May also be closed.

Signed-off-by: Nick Beirne <nick@tunnelbear.com>